Revert "Separated Login Page and functions From Main Page"
