+++ .gitignore
@@ -1,4 +1,3 @@
 node_modules
 .env
 .clj-kondo
-frontend/react-dev-server.log
