--- .q_file_picker/quick_open_recently_files.db
@@ -1,16 +1,0 @@
-# Auto generated, please do not modify# auto generated file, used to cache recently opened files
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\backend\src\user.js
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\frontend\src\App.js
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\backend\src\routes\ProtectedRoute.js
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\frontend\src\setupProxy.js
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\backend\src\server.js
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\frontend\src\pages\Main.js
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\frontend\src\pages\Gallery.js
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\frontend\src\AppContext.js
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\frontend\src\pages\ChatContainer.js
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\frontend\src\style\app.css
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\frontend\package.json
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\frontend\react-dev-server.log
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\backend\src\routes\MessagingRoute.js
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\backend\package.json
-d:\Andrea\Documents\GitHub\SOEN341-Project-Integration\frontend\node_modules\react-scripts\config\webpackDevServer.config.js
