--- backend/src/react-dev-server.log
@@ -1,1 +1,0 @@
-Route not found: POST /auth/loginRoute not found: POST /auth/loginRoute not found: POST /auth/loginRoute not found: POST /api/testRoute not found: POST /api/testRoute not found: POST /auth/loginRoute not found: POST /auth/loginRoute not found: POST /auth/loginRoute not found: POST /auth/loginRoute not found: POST /auth/loginRoute not found: POST /auth/loginRoute not found: POST /auth/loginRoute not found: POST /auth/loginRoute not found: GET /galleries?email=alice@gmail.comRoute not found: GET /galleries?UserID=undefinedRoute not found: GET /galleries?UserID=undefinedRoute not found: GET /galleries?UserID=undefinedRoute not found: GET /galleries?UserID=undefinedRoute not found: GET /galleries?UserID=undefinedRoute not found: GET /galleries?UserID=undefinedRoute not found: GET /galleries?UserID=nullRoute not found: GET /galleries?UserID=undefinedRoute not found: GET /galleries?UserID=undefinedRoute not found: GET /galleries?UserID=undefinedRoute not found: GET /galleries?UserID=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /galleries?UserID=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /galleries?UserID=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /galleries?UserID=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /galleries?UserID=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /galleries?UserID=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /user/galleries?UserID=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /user/galleriesRoute not found: GET /user/galleries?id=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /user/galleries?id=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /user/galleries?id=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /user/galleries?id=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /user/galleries?id=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /user/galleries?id=80dd60a3-96d1-4b48-a9ab-9c22f60b748fRoute not found: GET /user/galleries?id=80dd60a3-96d1-4b48-a9ab-9c22f60b748f
