+++ backend/src/routes/ProtectedRoute.js
@@ -1,11 +1,11 @@
 import express from "express";
+const router = express.Router();
 import { supabase } from "../server.js";
 
-const router = express.Router();
 router.use(express.json());
 
 // Registration route 
-router.post("/auth/register", async (req, res) => {
+router.post("/api/auth/register", async (req, res) => {
     const { email, password, username } = req.body;
 
     if (!email || !password ) {
@@ -46,28 +46,27 @@ router.post("/auth/register", async (req, res) => {
 });
 
 // Login route
-router.post("/auth/login", async (req, res) => {
+router.post("/api/auth/login", async (req, res) => {
     const { email, password } = req.body;
-    
+
     const { data, error } = await supabase.auth.signInWithPassword({
         email,
         password,
     });
-        
+
     if (error) {
-        console.error("Supabase authentication error:", error.message);
         return res.status(400).json({ msg: error.message });
     }
-    
+
     res.json({ msg: "Login successful", token: data.session.access_token });
 });
 
 // Get user info 
-router.get("/auth/me", async (req, res) => {
+router.get("/api/auth/me", async (req, res) => {
     const token = req.header("Authorization")?.split(" ")[1];
 
     if (!token) return res.status(401).json({ msg: "Unauthorized" });
-    
+
     const { data, error } = await supabase.auth.getUser(token);
 
     if (error) return res.status(401).json({ msg: "Invalid token" });
@@ -76,7 +75,7 @@ router.get("/auth/me", async (req, res) => {
 });
 
 // Logout route
-router.post("/auth/logout", async (req, res) => {
+router.post("/api/auth/logout", async (req, res) => {
     const { error } = await supabase.auth.signOut();
 
     if (error) return res.status(400).json({ msg: error.message });
