+++ frontend/package.json
@@ -9,7 +9,6 @@
     "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.3.3",
-    "http-proxy-middleware": "^3.0.3",
     "jquery": "^3.7.1",
     "lucide": "^0.476.0",
     "lucide-react": "^0.476.0",
@@ -17,7 +16,8 @@
     "react": "^19.0.0",
     "react-bootstrap": "^2.10.9",
     "react-dom": "^19.0.0",
-    "react-scripts": "^5.0.1",
+    "react-resizable": "^3.0.5",
+    "react-scripts": "5.0.1",
     "sass": "^1.85.0",
     "web-vitals": "^2.1.4"
   },
@@ -44,5 +44,6 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  }
+  },
+  "proxy": "http://localhost:4000"
 }
