+++ frontend/src/pages/Login.js
@@ -7,7 +7,7 @@ function Login({onLogin}){
     const [password, setPassword] = useState('');
 
     // Handle form submission
-    const handleSubmit = async (e) => {
+    const handleSubmit = (e) => {
         console.log("DEBUG: handleSubmit has been called.");
         e.preventDefault();
         onLogin(email, password);  // Call onLogin passed from the parent (App)
@@ -27,23 +27,23 @@ function Login({onLogin}){
     <div>
         {showLogin ? (
 
-            <div className="wrapper" id = "loginWrapper">
+            <div class="wrapper" id = "loginWrapper">
             <form id = "loginForm" onSubmit={handleSubmit}>
                 <h1>Login</h1>
-                <div className="input-box">
+                <div class="input-box">
                     <input type="text" placeholder="Email" value={email} onChange={(e) => setEmail(e.target.value)}required/>
-                    <i className='bx bxs-user'></i>
+                    <i class='bx bxs-user'></i>
                 </div>
-                <div className="input-box">
+                <div class="input-box">
                     <input type="text" placeholder="Password" value={password} onChange={(e) => setPassword(e.target.value)}required/>
-                    <i className='bx bxs-lock-alt' ></i>
+                    <i class='bx bxs-lock-alt' ></i>
                 </div>
-                <div className="text">
+                <div class="text">
                     <label><input type="checkbox"/>Remeber Me</label>
                     <a href="#">Forgot Password</a>
                 </div>
-                <button type="submit" className="button">Login</button>
-                <div className="register-link">
+                <button type="submit" class="button">Login</button>
+                <div class="register-link">
                     <p>Don't have an account? <a onClick={handleClick} id = "openRegister">Register</a></p>
                 </div>
             </form>
@@ -51,28 +51,28 @@ function Login({onLogin}){
 
         ):(
 
-            <div className="wrapper" id="registerWrapper" style={{display: 'block'}}>
+            <div class="wrapper" id="registerWrapper" style={{display: 'block'}}>
             <form id="registerForm">
-                <span className="close" id="closeRegister">&times;</span>
+                <span class="close" id="closeRegister">&times;</span>
                     <h1>Sign Up</h1>
-                    <div className="input-box">
+                    <div class="input-box">
                         <input type="text" placeholder="Username" required/>
-                        <i className='bx bxs-user'></i>
+                        <i class='bx bxs-user'></i>
                     </div>
-                    <div className="input-box">
+                    <div class="input-box">
                         <input type="email" placeholder="Email" required/>
-                        <i className='bx bxs-envelope'></i>
+                        <i class='bx bxs-envelope'></i>
                     </div>
-                    <div className="input-box">
+                    <div class="input-box">
                         <input type="password" placeholder="Create password" required/>
-                        <i className='bx bxs-lock-alt'></i>
+                        <i class='bx bxs-lock-alt'></i>
                     </div>
-                    <div className="input-box">
+                    <div class="input-box">
                         <input type="password" placeholder="Confirm password" required/>
-                        <i className='bx bxs-lock-alt'></i>
+                        <i class='bx bxs-lock-alt'></i>
                     </div>
-                    <button type="submit" className="button">Sign Up</button>
-                    <div className="login-link">
+                    <button type="submit" class="button">Sign Up</button>
+                    <div class="login-link">
                         <p>Already have an account? <a onClick={handleClick} id="backToLogin">Login</a></p>
                     </div>
                 </form>
